final_project/requirements.txt
asgiref==3.7.2
Django==4.0.4
django-dotenv==1.4.2
djangorestframework==3.14.0
gunicorn==21.2.0
mysqlclient==2.2.4
numpy==1.26.4
packaging==23.2
pandas==2.2.1
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
